Fix : Transform filter service into a generic service
We need to re-use the service into namespace & instance. To avoid code repetition, the filter service has been made generic 
Changes have also been made in workload to use the service without it being linked to workload
