+++ controller/lib/src/external_api/generic/mod.rs
@@ -1,0 +1,2 @@
+pub mod filter;
+pub mod model;
