+++ controller/lib/src/external_api/generic/model.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+pub enum FilterError {
+    OutOfRange,
+}
+
+#[derive(Deserialize, Serialize)]
+pub struct Pagination {
+    pub limit: u32,
+    pub offset: u32,
+}
