+++ controller/lib/src/external_api/mod.rs
@@ -1,2 +1,3 @@
+pub mod generic;
 pub mod interface;
 mod workload;
