+++ controller/lib/src/external_api/workload/mod.rs
@@ -1,4 +1,3 @@
 pub mod controller;
-pub mod filter;
 pub mod model;
 pub mod service;
